+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 * added precision critical
 * rowdy's vital force now deals precision damage
 * profane ascention now works corectly for dex and strength based characters
+* sword saint's perfect criticaly now correctl costs 2 points of pool instead of 1
 
 ## Version 1.3.3
 * fixed mounted maniac not triggering